# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: KSeF Monitor Cen - API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /price-alerts:
    get:
      operationId: listPriceAlerts
      tags: [price-alerts]
      summary: List price alerts
      responses:
        "200":
          description: List of price alerts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PriceAlert"
    post:
      operationId: createPriceAlert
      tags: [price-alerts]
      summary: Create a price alert
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePriceAlertBody"
      responses:
        "201":
          description: Created alert
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PriceAlert"
...
  /price-alerts/{id}:
    patch:
      operationId: updatePriceAlert
      tags: [price-alerts]
      summary: Update a price alert (toggle active, change threshold or supplier)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePriceAlertBody"
      responses:
        "200":
          description: Updated alert
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PriceAlert"
        "404":
          description: Not found
    delete:
      operationId: deletePriceAlert
      tags: [price-alerts]
      summary: Delete a price alert
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /price-alerts/{id}/dismiss:
    post:
      operationId: dismissPriceAlert
      tags: [price-alerts]
      summary: Dismiss a triggered alert occurrence
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DismissPriceAlertBody"
      responses:
        "201":
          description: Dismissed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DismissedAlert"
...
components:
  schemas:
...
    PriceAlert:
      type: object
      properties:
        id:
          type: integer
        productName:
          type: string
        supplierId:
          type: ["integer", "null"]
        supplierName:
          type: ["string", "null"]
        thresholdPercent:
          type: number
        isActive:
          type: boolean
        createdAt:
          type: string
        matchedProductId:
          type: ["integer", "null"]
          description: >-
            Produkt, do którego alert faktycznie się dopasował (po znormalizowanej
            nazwie). `null` = alert nie pasuje do żadnego produktu i nigdy się nie
            odpali — UI pokazuje wtedy ostrzeżenie. Zwracane tylko przez listę.
      required:
        - id
        - productName
        - thresholdPercent
        - isActive
        - createdAt
    CreatePriceAlertBody:
      type: object
      properties:
        productName:
          type: string
        supplierId:
          type: ["integer", "null"]
        thresholdPercent:
          type: number
      required:
        - productName
        - thresholdPercent
    UpdatePriceAlertBody:
      type: object
      properties:
        isActive:
          type: boolean
        thresholdPercent:
          type: number
        supplierId:
          type: ["integer", "null"]
    DismissPriceAlertBody:
      type: object
      properties:
        alertDate:
          type: string
        productName:
          type: string
        supplierName:
          type: ["string", "null"]
        currentPrice:
          type: number
        previousPrice:
          type: number
        changePercent:
          type: number
        thresholdPercent:
          type: number
      required:
        - alertDate
        - productName
        - currentPrice
        - previousPrice
        - changePercent
        - thresholdPercent
    DismissedAlert:
      type: object
      properties:
        id:
          type: integer
        alertId:
          type: ["integer", "null"]
        productName:
          type: string
        supplierName:
          type: ["string", "null"]
        alertDate:
          type: string
        currentPrice:
          type: number
        previousPrice:
          type: number
        changePercent:
          type: number
        thresholdPercent:
          type: number
        dismissedAt:
          type: string
      required:
        - id
        - productName
        - alertDate
        - currentPrice
        - previousPrice
        - changePercent
        - thresholdPercent
        - dismissedAt
...